+++ packages/next-auth/src/index.tsx
@@ -67,8 +67,8 @@
  * @module next-auth
  */
 
-import { Auth, setEnvDefaults as coreSetEnvDefaults } from "@auth/core"
-import { reqWithEnvURL } from "./lib/env.js"
+import { Auth } from "@auth/core"
+import { actionWithEnv, reqWithEnvURL, setEnvDefaults } from "./lib/env.js"
 import { initAuth } from "./lib/index.js"
 import { signIn, signOut, update } from "./lib/actions.js"
 
@@ -372,27 +372,13 @@ export default function NextAuth(
     return {
       handlers: { GET: httpHandler, POST: httpHandler } as const,
       // @ts-expect-error
-      auth: initAuth(config, (c) => setEnvDefaults(c)),
-
-      signIn: (provider, options, authorizationParams) => {
-        const _config = config(undefined)
-        setEnvDefaults(_config)
-        return signIn(provider, options, authorizationParams, _config)
-      },
-      signOut: (options) => {
-        const _config = config(undefined)
-        setEnvDefaults(_config)
-        return signOut(options, _config)
-      },
-      unstable_update: (data) => {
-        const _config = config(undefined)
-        setEnvDefaults(_config)
-        return update(data, _config)
-      },
+      auth: initAuth(config, setEnvDefaults),
+      signIn: actionWithEnv(signIn, config),
+      signOut: actionWithEnv(signOut, config),
+      unstable_update: actionWithEnv(update, config),
     }
   }
   setEnvDefaults(config)
-  config.basePath ??= "/api/auth"
   const httpHandler = (req: NextRequest) => Auth(reqWithEnvURL(req), config)
   return {
     handlers: { GET: httpHandler, POST: httpHandler } as const,
@@ -407,24 +393,5 @@ export default function NextAuth(
     unstable_update: (data) => {
       return update(data, config)
     },
-  }
-}
-
-// this is the one that next-auth should be using everywhere
-/**
- * @internal
- * For backwards compatibility, `next-auth` checks for `NEXTAUTH_URL`
- * and the `basePath` by default is `/api/auth` instead of `/auth`
- * (which is the default for all other Auth.js integrations).
- */
-function setEnvDefaults(config: NextAuthConfig) {
-  try {
-    const url = process.env.AUTH_URL ?? process.env.NEXTAUTH_URL
-    // @ts-expect-error if empty, we expect it to throw and set in finally
-    config.basePath ||= new URL(url).pathname
-  } catch {
-  } finally {
-    config.basePath ||= "/api/auth"
-    coreSetEnvDefaults(process.env, config)
   }
 }
